requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
supabase==2.0.2
python-dotenv==1.0.0
Pillow>=9.0.0
imagehash==4.3.1
opencv-python>=4.8.0
numpy>=1.21.0 